server/requirements.txt
Flask==1.1.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==2.0.2
numpy>=1.20.3
scipy>=1.7.1
scikit-learn>=0.24.2
matplotlib>=3.4.3
pandas>=1.3.4
tensorflow-cpu==2.8.0
json==0.9.6

